Move test_false_4d
