+++ tests/object.rs
@@ -187,31 +187,3 @@ fn f32_nii_gz() {
     assert_eq!(volume.get_f32(&[5, 0, 4]).unwrap(), 0.4);
     assert_eq!(volume.get_f32(&[0, 8, 5]).unwrap(), 0.8);
 }
-
-#[cfg(feature = "ndarray_volumes")]
-#[test]
-fn test_false_4d() {
-    use ndarray::Ix3;
-    use nifti:: {InMemNiftiVolume, IntoNdArray};
-
-    let (w, h, d) = (5, 5, 5);
-    let mut header = NiftiHeader {
-        dim: [4, w, h, d, 1, 1, 1, 1],
-        datatype: 2,
-        bitpix: 8,
-        ..Default::default()
-    };
-    let raw_data = vec![0; (w * h * d) as usize];
-    let mut volume = InMemNiftiVolume::from_raw_data(&header, raw_data).unwrap();
-    assert_eq!(header.dim[0], 4);
-    assert_eq!(volume.dimensionality(), 4);
-    if header.dim[header.dim[0] as usize] == 1 {
-        header.dim[0] -= 1;
-        volume = InMemNiftiVolume::from_raw_data(&header, volume.to_raw_data()).unwrap();
-    }
-    assert_eq!(volume.dimensionality(), 3);
-    let dyn_data = volume.to_ndarray::<f32>().unwrap();
-    assert_eq!(dyn_data.ndim(), 3);
-    let data = dyn_data.into_dimensionality::<Ix3>().unwrap();
-    assert_eq!(data.ndim(), 3); // Obvious, but it's to avoid being optimized away
-}
